Added confirmation popup and removed unnecessary button

+++ kemaskinipekerja/index.php
@@ -55,7 +55,7 @@
                                 <a href="../kemaskinipekerjafun/index.php?index=<?php echo $info['idpekerja'];?>">
                                     <button class="zi-btn action-btn">KEMASKINI</button>
                                 </a>
-                                <a href="../functions/hapuspekerja.php?index=<?php echo $info['idpekerja'];?>">
+                                <a href="../functions/hapuspekerja.php?index=<?php echo $info['idpekerja'];?>" onclick="return confirm('Hapus maklumat pekerja dari sistem?');">
                                     <button class="zi-btn action-btn">HAPUS</button>
                                 </a>
                             </td>
